Move connection test a layer down, refactor create_connected_connections testing function
