+++ libvcx/src/utils/devsetup.rs
@@ -446,7 +446,7 @@ mod tests {
         let mut institution = Faber::setup();
         let mut consumer1 = Alice::setup();
 
-        let (_faber, _alice) = connection::tests::create_connected_connections(&mut consumer1, &mut institution);
-        let (_faber, _alice) = connection::tests::create_connected_connections(&mut consumer1, &mut institution);
+        let (_faber, _alice) = connection::tests::create_and_store_connected_connections(&mut consumer1, &mut institution);
+        let (_faber, _alice) = connection::tests::create_and_store_connected_connections(&mut consumer1, &mut institution);
     }
 }
